refactor: extract `IntervalSet<T>` into core
This will end up being a useful data structure beyond the previous version of the simulator.
I also want to modify it for faster insertion and better locality using a vector.

This commit focuses on extracting the interval_set from the previous simulator and porting its tests.
